requirements/local.txt
# Extend base.txt
-r base.txt

Werkzeug  # pyup: # https://github.com/pallets/werkzeug
psycopg2-binary>=2.8.3  # https://github.com/psycopg/psycopg2


# Code quality
# ------------------------------------------------------------------------------
flake8
black
pylint-django
isort
#pre-commit>=1.20.0  # https://github.com/pre-commit/pre-commit


django-debug-toolbar
django-extensions